create useScannedBoxesActions hook
